feat: change paths for endpoints
