+++ web/swagger/api.yaml
@@ -308,12 +308,37 @@ paths:
       summary: Get markets information for the DEX.
       tags:
       - Markets
-  /v0/orders/cancel:
-    post:
+  /v0/order-book/{market-id}:
+    get:
+      description: Get order book for a specific market.
+      parameters:
+      - in: path
+        name: market-id
+        required: true
+        type: string
+      - format: bech32
+        in: query
+        name: address
+        required: false
+        type: string
+      produces:
+      - application/json;charset=utf-8
+      responses:
+        '200':
+          description: ''
+          schema:
+            $ref: '#/definitions/OrderBookInfo'
+        '400':
+          description: Invalid `address` or `market-id`
+      summary: Order book
+      tags:
+      - Order Book
+  /v0/orders:
+    delete:
       consumes:
       - application/json;charset=utf-8
-      description: Cancel order(s). This differs from earlier endpoint in that it
-        would also sign & submit the built transaction
+      description: Cancel order(s). This endpoint would also sign & submit the built
+        transaction
       parameters:
       - in: body
         name: body
@@ -332,7 +357,30 @@ paths:
       summary: Cancel order(s)
       tags:
       - Orders
-  /v0/orders/cancel/tx/generate:
+    post:
+      consumes:
+      - application/json;charset=utf-8
+      description: Create an order. This endpoint would also sign & submit the built
+        transaction
+      parameters:
+      - in: body
+        name: body
+        required: true
+        schema:
+          $ref: '#/definitions/PlaceOrderParameters'
+      produces:
+      - application/json;charset=utf-8
+      responses:
+        '200':
+          description: ''
+          schema:
+            $ref: '#/definitions/PlaceOrderTransactionDetails'
+        '400':
+          description: Invalid `body`
+      summary: Create an order
+      tags:
+      - Orders
+  /v0/orders/tx/build-cancel:
     post:
       consumes:
       - application/json;charset=utf-8
@@ -355,31 +403,7 @@ paths:
       summary: Build transaction to cancel order(s)
       tags:
       - Orders
-  /v0/orders/open:
-    post:
-      consumes:
-      - application/json;charset=utf-8
-      description: Create an order. This differs from earlier endpoint in that it
-        would also sign & submit the built transaction
-      parameters:
-      - in: body
-        name: body
-        required: true
-        schema:
-          $ref: '#/definitions/PlaceOrderParameters'
-      produces:
-      - application/json;charset=utf-8
-      responses:
-        '200':
-          description: ''
-          schema:
-            $ref: '#/definitions/PlaceOrderTransactionDetails'
-        '400':
-          description: Invalid `body`
-      summary: Create an order
-      tags:
-      - Orders
-  /v0/orders/open/tx/generate:
+  /v0/orders/tx/build-open:
     post:
       consumes:
       - application/json;charset=utf-8
@@ -402,31 +426,6 @@ paths:
       summary: Build transaction to create order
       tags:
       - Orders
-  /v0/orders/{market_id}:
-    get:
-      description: Get on-chain order(s)
-      parameters:
-      - in: path
-        name: market_id
-        required: true
-        type: string
-      - format: bech32
-        in: query
-        name: address
-        required: false
-        type: string
-      produces:
-      - application/json;charset=utf-8
-      responses:
-        '200':
-          description: ''
-          schema:
-            $ref: '#/definitions/OrderBookInfo'
-        '400':
-          description: Invalid `address` or `market_id`
-      summary: Get order(s)
-      tags:
-      - Orders
   /v0/settings:
     get:
       description: Get server settings such as network, version, and revision
@@ -440,7 +439,7 @@ paths:
       summary: Server settings
       tags:
       - Settings
-  /v0/trading_fees:
+  /v0/trading-fees:
     get:
       description: Get trading fees of DEX.
       produces:
@@ -476,7 +475,7 @@ paths:
       summary: Sign a transaction
       tags:
       - Transaction
-  /v0/tx/sign_and_submit:
+  /v0/tx/sign-and-submit:
     post:
       consumes:
       - application/json;charset=utf-8
@@ -537,3 +536,5 @@ tags:
   name: Trading Fees
 - description: Endpoint to fetch asset details.
   name: Assets
+- description: Endpoint to fetch order book.
+  name: Order Book
